Refactor: refactor for version 1
